+++ packages/crepe/src/core/index.ts
@@ -1,7 +1,3 @@
 export { Crepe, type CrepeConfig } from './crepe'
-export {
-  CrepeBuilder,
-  type CrepeBuilderConfig,
-  crepeCtx,
-  FeaturesCtx,
-} from './builder'
+export { CrepeBuilder, type CrepeBuilderConfig } from './builder'
+export { useCrepe, useCrepeFeatures } from './slice'
