+++ packages/crepe/src/feature/code-mirror/index.ts
@@ -11,9 +11,9 @@ import { basicSetup } from 'codemirror'
 
 import type { DefineFeature, Icon } from '../shared'
 
+import { crepeFeatureConfig } from '../../core/slice'
 import { chevronDownIcon, clearIcon, editIcon, searchIcon } from '../../icons'
 import { visibilityOffIcon } from '../../icons/visibility-off'
-import { crepeFeatureConfig } from '../../utils'
 import { CrepeFeature } from '../index'
 
 interface CodeMirrorConfig {
