+++ packages/crepe/src/feature/link-tooltip/index.ts
@@ -6,8 +6,8 @@ import {
 
 import type { DefineFeature, Icon } from '../shared'
 
+import { crepeFeatureConfig } from '../../core/slice'
 import { copyIcon, editIcon, removeIcon, confirmIcon } from '../../icons'
-import { crepeFeatureConfig } from '../../utils'
 import { CrepeFeature } from '../index'
 
 interface LinkTooltipConfig {
