+++ packages/crepe/src/feature/list-item/index.ts
@@ -7,12 +7,12 @@ import {
 
 import type { DefineFeature, Icon } from '../shared'
 
+import { crepeFeatureConfig } from '../../core/slice'
 import {
   bulletIcon,
   checkBoxCheckedIcon,
   checkBoxUncheckedIcon,
 } from '../../icons'
-import { crepeFeatureConfig } from '../../utils'
 import { CrepeFeature } from '../index'
 
 export interface ListItemConfig {
