+++ packages/crepe/src/index.ts
@@ -1,9 +1,2 @@
 export { CrepeFeature } from './feature'
-export {
-  Crepe,
-  type CrepeConfig,
-  crepeCtx,
-  FeaturesCtx,
-  type CrepeBuilderConfig,
-  CrepeBuilder,
-} from './core'
+export * from './core'
